Create README.md
Hi, I'm Don McLean, Professor of Music at the University of Toronto, and former Dean of Music 2011-2021, and prior to that Dean of the Schulich School of Music of McGill University 2011-2010. I have been working in higher education in Music since 1979 and am taking a Boot Camp coding course while on leave 2021-2022.